+++ xbird-cloud-netflix/xbird-cloud-netflix-turbine/src/main/java/com/xbird/cloud/netflix/turbine/TurbineApplication.java
@@ -1,0 +1,20 @@
+package com.xbird.cloud.netflix.turbine;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.netflix.hystrix.EnableHystrix;
+import org.springframework.cloud.netflix.hystrix.dashboard.EnableHystrixDashboard;
+import org.springframework.cloud.netflix.turbine.EnableTurbine;
+
+@SpringBootApplication
+@EnableEurekaClient
+@EnableTurbine
+@EnableHystrix
+@EnableHystrixDashboard
+public class TurbineApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(TurbineApplication.class, args);
+  }
+}